Move 3DS2 tests to their own file. GIT_VALID_PII_OVERRIDE
