+++ paymentsheet-example/src/androidTest/java/com/stripe/android/lpm/TestCard.kt
@@ -15,7 +15,6 @@ import com.stripe.android.paymentsheet.example.playground.settings.DefaultBillin
 import com.stripe.android.paymentsheet.example.playground.settings.DefaultBillingAddressSettingsDefinition
 import com.stripe.android.paymentsheet.example.samples.ui.shared.PAYMENT_METHOD_SELECTOR_TEST_TAG
 import com.stripe.android.paymentsheet.ui.SAVED_PAYMENT_OPTION_TEST_TAG
-import com.stripe.android.test.core.AuthorizeAction
 import com.stripe.android.test.core.FieldPopulator
 import com.stripe.android.test.core.TestParameters
 import org.junit.Test
@@ -38,71 +37,6 @@ internal class TestCard : BasePlaygroundTest() {
             populateCustomLpmFields = {
                 populateCardDetails()
             },
-        )
-    }
-
-    @Test
-    fun test3DS2HSBCHTML() {
-        testDriver.confirmNewOrGuestComplete(
-            testParameters.copy(
-                authorizationAction = AuthorizeAction.Test3DS2.HSBCHTML,
-            ),
-            populateCustomLpmFields = {
-                populateCardDetails()
-            },
-            values = FieldPopulator.Values(cardNumber = "4000582600000292"),
-        )
-    }
-
-    @Test
-    fun test3DS2OTP() {
-        testDriver.confirmNewOrGuestComplete(
-            testParameters.copy(
-                authorizationAction = AuthorizeAction.Test3DS2.OTP,
-            ),
-            populateCustomLpmFields = {
-                populateCardDetails()
-            },
-            values = FieldPopulator.Values(cardNumber = "4000582600000045"),
-        )
-    }
-
-    @Test
-    fun test3DS2OOB() {
-        testDriver.confirmNewOrGuestComplete(
-            testParameters.copy(
-                authorizationAction = AuthorizeAction.Test3DS2.OOB,
-            ),
-            populateCustomLpmFields = {
-                populateCardDetails()
-            },
-            values = FieldPopulator.Values(cardNumber = "4000582600000094"),
-        )
-    }
-
-    @Test
-    fun test3DS2SingleSelect() {
-        testDriver.confirmNewOrGuestComplete(
-            testParameters.copy(
-                authorizationAction = AuthorizeAction.Test3DS2.SingleSelect,
-            ),
-            populateCustomLpmFields = {
-                populateCardDetails()
-            },
-            values = FieldPopulator.Values(cardNumber = "4000582600000102"),
-        )
-    }
-
-    @Test
-    fun test3DS2MultiSelect() {
-        testDriver.confirmNewOrGuestComplete(
-            testParameters.copy(
-                authorizationAction = AuthorizeAction.Test3DS2.MultiSelect,
-            ),
-            populateCustomLpmFields = {
-                populateCardDetails()
-            },
-            values = FieldPopulator.Values(cardNumber = "4000582600000110"),
         )
     }
 
@@ -331,19 +265,6 @@ internal class TestCard : BasePlaygroundTest() {
                     timeoutMillis = 5000L
                 )
             },
-        )
-    }
-
-    @Test
-    fun testCardWith3ds2() {
-        testDriver.confirmNewOrGuestComplete(
-            testParameters.copy(
-                authorizationAction = AuthorizeAction.Authorize3ds2,
-            ),
-            populateCustomLpmFields = {
-                populateCardDetails()
-            },
-            values = FieldPopulator.Values(cardNumber = "4000000000003220")
         )
     }
